Update to use new vhost syntax
https://github.com/googleapis/google-cloud-java/pull/6183 was merged
into google-cloud-java, and it changed some of the syntax (e.g. the
method name for virtual hosted urls changed, and explicitly specifying
the virtual hostname is no longer possible with that method). The syntax
updates have been reflected in the main-module examples.

This also updates the google-cloud-java submodule to use the merged
"master" branch ("master" of my fork is up-to-date with "master" of the
googleapis upstream as of today), along with changing main module to
depend on the newly-updated version google-cloud-storage
(1.93.1-SNAPSHOT).